requirements_PSS_Base.txt
# platform: debian
python=3.7.*
pip=22.1.2
numpy=1.20.1
networkx=2.6.3
pefile=2019.4.18
pytorch=1.12.1
